+++ tests/test_stream.py
@@ -326,7 +326,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         iq = make_test_iq(type_=structs.IQType.GET)
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         obj = run_coroutine(self.sent_stanzas.get())
         self.assertIs(obj, iq)
@@ -336,7 +336,7 @@ class TestStanzaStream(StanzaStreamTestBase):
     def test_enqueue_validates_stanza(self):
         iq = unittest.mock.Mock()
 
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         iq.validate.assert_called_with()
 
@@ -1330,14 +1330,14 @@ class TestStanzaStream(StanzaStreamTestBase):
         self.stream.on_failure.connect(failure_handler)
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         iq_sent = run_coroutine(self.sent_stanzas.get())
         self.assertIs(iq, iq_sent)
 
         self.xmlstream.send_xso = unittest.mock.MagicMock(
             side_effect=RuntimeError())
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
         self.stream.recv_stanza(iq)
         self.stream.stop()
 
@@ -1354,14 +1354,14 @@ class TestStanzaStream(StanzaStreamTestBase):
         self.stream.on_failure.connect(failure_handler)
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         iq_sent = run_coroutine(self.sent_stanzas.get())
         self.assertIs(iq, iq_sent)
 
         self.xmlstream.send_xso = unittest.mock.MagicMock(
             side_effect=RuntimeError())
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
         self.stream.recv_stanza(iq)
         self.stream.stop()
 
@@ -1561,7 +1561,7 @@ class TestStanzaStream(StanzaStreamTestBase):
             run_coroutine(asyncio.sleep(0))
             self.assertTrue(self.stream.running)
 
-            token = self.stream.enqueue(make_test_message())
+            token = self.stream._enqueue(make_test_message())
 
             run_coroutine(self.stream.close())
 
@@ -1579,7 +1579,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 "get",
                 new=get_mock):
 
-            token = self.stream.enqueue(make_test_message())
+            token = self.stream._enqueue(make_test_message())
 
             run_coroutine(self.stream.close())
 
@@ -1591,7 +1591,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         run_coroutine(self.stream.close())
 
         with self.assertRaisesRegex(ConnectionError, r"close\(\) called"):
-            self.stream.enqueue(unittest.mock.sentinel.stanza)
+            self.stream._enqueue(unittest.mock.sentinel.stanza)
 
     def test_enqueue_works_after_close_and_start(self):
         run_coroutine(self.stream.close())
@@ -1599,7 +1599,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         iq = make_test_iq(type_=structs.IQType.GET)
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         obj = run_coroutine(self.sent_stanzas.get())
         self.assertIs(obj, iq)
@@ -1621,7 +1621,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         self.xmlstream.send_xso = send_handler
 
         tokens = [
-            self.stream.enqueue(
+            self.stream._enqueue(
                 iq,
                 on_state_change=state_change_handler)
             for iq in iqs
@@ -1782,7 +1782,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         )
 
     def test_enqueue_returns_token(self):
-        token = self.stream.enqueue(make_test_iq())
+        token = self.stream._enqueue(make_test_iq())
         self.assertIsInstance(
             token,
             stream.StanzaToken)
@@ -1790,7 +1790,7 @@ class TestStanzaStream(StanzaStreamTestBase):
     def test_abort_stanza(self):
         iqs = [make_test_iq() for i in range(3)]
         self.stream.start(self.xmlstream)
-        tokens = [self.stream.enqueue(iq) for iq in iqs]
+        tokens = [self.stream._enqueue(iq) for iq in iqs]
         tokens[1].abort()
         run_coroutine(asyncio.sleep(0))
 
@@ -1917,7 +1917,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaises(ConnectionError):
             run_coroutine(asyncio.gather(
                 kill_it(),
-                self.stream.send_iq_and_wait_for_reply(iq)
+                self.stream._send_immediately(iq)
             ))
 
     def _test_inbound_presence_filter(self, filter_attr, **register_kwargs):
@@ -2129,7 +2129,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         filter_attr.register(filter_func, **register_kwargs)
 
         self.stream.start(self.xmlstream)
-        token = self.stream.enqueue(pres)
+        token = self.stream._enqueue(pres)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2153,7 +2153,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         filter_func.reset_mock()
         filter_func.return_value = None
 
-        token = self.stream.enqueue(pres)
+        token = self.stream._enqueue(pres)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2205,7 +2205,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         )
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(pres)
+        self.stream._enqueue(pres)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2228,7 +2228,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         filter_attr.register(filter_func, **register_kwargs)
 
         self.stream.start(self.xmlstream)
-        token = self.stream.enqueue(msg)
+        token = self.stream._enqueue(msg)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2252,7 +2252,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         filter_func.reset_mock()
         filter_func.return_value = None
 
-        token = self.stream.enqueue(msg)
+        token = self.stream._enqueue(msg)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2304,7 +2304,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         )
 
         self.stream.start(self.xmlstream)
-        self.stream.enqueue(msg)
+        self.stream._enqueue(msg)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -2330,7 +2330,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         msg = make_test_message()
         msg.autoset_id()
 
-        self.stream.enqueue(msg)
+        self.stream._enqueue(msg)
 
         self.xmlstream.on_closing(None)
 
@@ -2879,21 +2879,21 @@ class TestStanzaStream(StanzaStreamTestBase):
         iq = make_test_iq()
 
         base = unittest.mock.Mock()
-        base.enqueue = CoroutineMock()
-        base.enqueue.return_value = \
+        base._enqueue = CoroutineMock()
+        base._enqueue.return_value = \
             unittest.mock.sentinel.token_await_result
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             run_coroutine(self.stream.send_and_wait_for_sent(
                 iq
             ))
 
-        base.enqueue.assert_called_with(unittest.mock.ANY)
+        base._enqueue.assert_called_with(unittest.mock.ANY)
 
     def test_send_and_wait_for_sent_emits_deprecation_warning(self):
         iq = make_test_iq()
@@ -2909,14 +2909,14 @@ class TestStanzaStream(StanzaStreamTestBase):
         pres = make_test_presence()
 
         base = unittest.mock.Mock()
-        base.enqueue = CoroutineMock()
-        base.enqueue.return_value = \
+        base._enqueue = CoroutineMock()
+        base._enqueue.return_value = \
             unittest.mock.sentinel.token_await_result
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -2925,23 +2925,23 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.register_iq_response_future,
             ))
 
-            run_coroutine(self.stream.send(pres))
+            run_coroutine(self.stream._send_immediately(pres))
 
         base.register_iq_response_future.assert_not_called()
-        base.enqueue.assert_called_with(unittest.mock.ANY)
+        base._enqueue.assert_called_with(unittest.mock.ANY)
 
     def test_send_awaits_stanza_token_for_message(self):
         message = make_test_presence()
 
         base = unittest.mock.Mock()
-        base.enqueue = CoroutineMock()
-        base.enqueue.return_value = \
+        base._enqueue = CoroutineMock()
+        base._enqueue.return_value = \
             unittest.mock.sentinel.token_await_result
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -2950,23 +2950,23 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.register_iq_response_future,
             ))
 
-            run_coroutine(self.stream.send(message))
+            run_coroutine(self.stream._send_immediately(message))
 
         base.register_iq_response_future.assert_not_called()
-        base.enqueue.assert_called_with(unittest.mock.ANY)
+        base._enqueue.assert_called_with(unittest.mock.ANY)
 
     def test_send_awaits_stanza_token_for_iq_response(self):
         iq = make_test_iq(type_=aioxmpp.IQType.RESULT)
 
         base = unittest.mock.Mock()
-        base.enqueue = CoroutineMock()
-        base.enqueue.return_value = \
+        base._enqueue = CoroutineMock()
+        base._enqueue.return_value = \
             unittest.mock.sentinel.token_await_result
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -2975,10 +2975,10 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.register_iq_response_future,
             ))
 
-            run_coroutine(self.stream.send(iq))
+            run_coroutine(self.stream._send_immediately(iq))
 
         base.register_iq_response_future.assert_not_called()
-        base.enqueue.assert_called_with(unittest.mock.ANY)
+        base._enqueue.assert_called_with(unittest.mock.ANY)
 
     def test_send_awaits_stanza_token_for_iq_and_registers_for_reply(self):
         iq = make_test_iq()
@@ -2988,12 +2988,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut = asyncio.Future()
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3002,11 +3002,11 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq))
+            task = asyncio.async(self.stream._send_immediately(iq))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3035,12 +3035,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut = asyncio.Future()
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3049,11 +3049,11 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq))
+            task = asyncio.async(self.stream._send_immediately(iq))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3088,12 +3088,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut = asyncio.Future()
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3102,11 +3102,11 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq))
+            task = asyncio.async(self.stream._send_immediately(iq))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3136,19 +3136,21 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut = asyncio.Future()
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
-            task = asyncio.async(self.stream.send(iq, timeout=0.001))
+            task = asyncio.async(self.stream._send_immediately(
+                iq,
+                timeout=0.001))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
 
             stanza_fut.set_result(None)
 
@@ -3162,12 +3164,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut = asyncio.Future()
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3176,11 +3178,13 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq, timeout=0.001))
+            task = asyncio.async(self.stream._send_immediately(
+                iq,
+                timeout=0.001))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3207,12 +3211,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut.set_result(None)
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3221,11 +3225,11 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq))
+            task = asyncio.async(self.stream._send_immediately(iq))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3249,12 +3253,12 @@ class TestStanzaStream(StanzaStreamTestBase):
         stanza_fut.set_result(None)
 
         base = unittest.mock.Mock()
-        base.enqueue.return_value = stanza_fut
+        base._enqueue.return_value = stanza_fut
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch.object(
                 self.stream,
-                "enqueue",
-                new=base.enqueue
+                "_enqueue",
+                new=base._enqueue
             ))
 
             stack.enter_context(unittest.mock.patch.object(
@@ -3263,11 +3267,11 @@ class TestStanzaStream(StanzaStreamTestBase):
                 new=base.iq_response_map,
             ))
 
-            task = asyncio.async(self.stream.send(iq))
+            task = asyncio.async(self.stream._send_immediately(iq))
             run_coroutine(asyncio.sleep(0.01))
 
             self.assertFalse(task.done())
-            base.enqueue.assert_called_with(unittest.mock.ANY)
+            base._enqueue.assert_called_with(unittest.mock.ANY)
             base.iq_response_map.add_listener.assert_called_once_with(
                 (iq.to, iq.id_),
                 unittest.mock.ANY,
@@ -3297,7 +3301,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         self.xmlstream.send_xso.side_effect = exc
 
         self.stream.start(self.xmlstream)
-        token = self.stream.enqueue(msg)
+        token = self.stream._enqueue(msg)
         self.assertEqual(token.state, stream.StanzaState.ACTIVE)
 
         run_coroutine(asyncio.sleep(0.05))
@@ -3444,7 +3448,9 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaisesRegex(
                 ValueError,
                 r"cb not supported with non-IQ non-request stanzas"):
-            run_coroutine(self.stream.send(msg, cb=unittest.mock.sentinel.cb))
+            run_coroutine(self.stream._send_immediately(
+                msg,
+                cb=unittest.mock.sentinel.cb))
 
         self.assertTrue(self.sent_stanzas.empty())
 
@@ -3456,7 +3462,9 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaisesRegex(
                 ValueError,
                 r"cb not supported with non-IQ non-request stanzas"):
-            run_coroutine(self.stream.send(pres, cb=unittest.mock.sentinel.cb))
+            run_coroutine(self.stream._send_immediately(
+                pres,
+                cb=unittest.mock.sentinel.cb))
 
         self.assertTrue(self.sent_stanzas.empty())
 
@@ -3468,7 +3476,9 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaisesRegex(
                 ValueError,
                 r"cb not supported with non-IQ non-request stanzas"):
-            run_coroutine(self.stream.send(iq, cb=unittest.mock.sentinel.cb))
+            run_coroutine(self.stream._send_immediately(
+                iq,
+                cb=unittest.mock.sentinel.cb))
 
         self.assertTrue(self.sent_stanzas.empty())
 
@@ -3481,7 +3491,7 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.start(self.xmlstream)
 
-        task = asyncio.async(self.stream.send(iq, cb=cb))
+        task = asyncio.async(self.stream._send_immediately(iq, cb=cb))
 
         run_coroutine(self.sent_stanzas.get())
         self.assertFalse(task.done())
@@ -3519,7 +3529,7 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.start(self.xmlstream)
 
-        task = asyncio.async(self.stream.send(iq, cb=cb))
+        task = asyncio.async(self.stream._send_immediately(iq, cb=cb))
 
         run_coroutine(self.sent_stanzas.get())
         self.assertFalse(task.done())
@@ -3553,7 +3563,7 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.start(self.xmlstream)
 
-        task = asyncio.async(self.stream.send(iq, cb=cb))
+        task = asyncio.async(self.stream._send_immediately(iq, cb=cb))
 
         run_coroutine(self.sent_stanzas.get())
         self.assertFalse(task.done())
@@ -3580,7 +3590,7 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.start(self.xmlstream)
 
-        task = asyncio.async(self.stream.send(iq, cb=cb))
+        task = asyncio.async(self.stream._send_immediately(iq, cb=cb))
 
         run_coroutine(self.sent_stanzas.get())
         self.assertFalse(task.done())
@@ -3613,7 +3623,7 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.start(self.xmlstream)
 
-        task = asyncio.async(self.stream.send(iq, cb=cb))
+        task = asyncio.async(self.stream._send_immediately(iq, cb=cb))
 
         run_coroutine(self.sent_stanzas.get())
         self.assertFalse(task.done())
@@ -4009,7 +4019,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         @asyncio.coroutine
         def starter():
             sm_start_future = asyncio.async(self.stream.start_sm())
-            self.stream.enqueue(iq_sent)
+            self.stream._enqueue(iq_sent)
 
         self.stream.start(self.xmlstream)
         run_coroutine_with_peer(
@@ -4065,7 +4075,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         tokens = [
-            self.stream.enqueue(
+            self.stream._enqueue(
                 iq, on_state_change=state_change_handler)
             for iq in iqs]
 
@@ -4163,7 +4173,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         tokens = [
-            self.stream.enqueue(
+            self.stream._enqueue(
                 iq, on_state_change=state_change_handler)
             for iq in iqs]
 
@@ -4352,7 +4362,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         for iq in iqs:
-            self.stream.enqueue(iq)
+            self.stream._enqueue(iq)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -4379,7 +4389,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
 
         # enqueue a stanza before resumption and check that the sequence is
         # correct (resumption-generated stanzas before new stanzas)
-        self.stream.enqueue(additional_iq)
+        self.stream._enqueue(additional_iq)
 
         run_coroutine_with_peer(
             self.stream.resume_sm(self.xmlstream),
@@ -4427,7 +4437,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         for iq in iqs:
-            self.stream.enqueue(iq)
+            self.stream._enqueue(iq)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -4455,7 +4465,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
 
         # enqueue a stanza before resumption and check that the sequence is
         # correct (resumption-generated stanzas before new stanzas)
-        self.stream.enqueue(additional_iq)
+        self.stream._enqueue(additional_iq)
 
         run_coroutine_with_peer(
             self.stream.resume_sm(self.xmlstream),
@@ -4503,7 +4513,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         for iq in iqs:
-            self.stream.enqueue(iq)
+            self.stream._enqueue(iq)
 
         run_coroutine(asyncio.sleep(0))
 
@@ -4530,7 +4540,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
 
         # enqueue a stanza before resumption and check that the sequence is
         # correct (resumption-generated stanzas before new stanzas)
-        self.stream.enqueue(additional_iq)
+        self.stream._enqueue(additional_iq)
 
         run_coroutine_with_peer(
             self.stream.resume_sm(self.xmlstream),
@@ -4697,7 +4707,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         iq = make_test_iq()
-        self.stream.enqueue(iq)
+        self.stream._enqueue(iq)
 
         run_coroutine(self.xmlstream.run_test([
             XMLStreamMock.Send(iq),
@@ -4734,9 +4744,9 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
         )
 
         run_coroutine(asyncio.sleep(0))
-        self.stream.enqueue(iqs[0])
+        self.stream._enqueue(iqs[0])
         run_coroutine(asyncio.sleep(0.05))
-        self.stream.enqueue(iqs[1])
+        self.stream._enqueue(iqs[1])
         run_coroutine(asyncio.sleep(0.06))
 
         run_coroutine(self.xmlstream.run_test([
@@ -4772,12 +4782,12 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
             self.xmlstream.run_test(self.successful_sm)
         )
 
-        self.stream.enqueue(iqs[0])
+        self.stream._enqueue(iqs[0])
         run_coroutine(self.xmlstream.run_test([
             XMLStreamMock.Send(iqs[0]),
             XMLStreamMock.Send(nonza.SMRequest()),
         ]))
-        self.stream.enqueue(iqs[1])
+        self.stream._enqueue(iqs[1])
         run_coroutine(self.xmlstream.run_test([
             XMLStreamMock.Send(iqs[1]),
             XMLStreamMock.Send(
@@ -4870,7 +4880,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
             self.xmlstream.run_test(self.successful_sm)
         )
 
-        tokens = [self.stream.enqueue(iq) for iq in iqs]
+        tokens = [self.stream._enqueue(iq) for iq in iqs]
 
         run_coroutine(self.xmlstream.run_test([
             XMLStreamMock.Send(iqs[0]),
@@ -5101,7 +5111,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
 
         self.stream.stop()
 
-        self.stream.enqueue(pres)
+        self.stream._enqueue(pres)
 
         self.assertIs(
             pres,
@@ -5118,7 +5128,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
             self.xmlstream.run_test(self.successful_sm)
         )
 
-        self.stream.enqueue(pres)
+        self.stream._enqueue(pres)
         self.stream.stop()
 
         self.assertIs(
@@ -5136,7 +5146,7 @@ class TestStanzaStreamSM(StanzaStreamTestBase):
             self.xmlstream.run_test(self.successful_sm)
         )
 
-        token = self.stream.enqueue(pres)
+        token = self.stream._enqueue(pres)
 
         run_coroutine_with_peer(
             self.stream.close(),